refactor(rs): consolidate remote sensing pipeline into deep SpectralRasterEngine and delete shallow rs_service
