+++ app/services/spatial_tasks.py
@@ -17,7 +17,7 @@ from shapely.geometry import box, mapping
 from app.services.task_queue import celery_app
 from app.services.spatial_analyzer import SpatialAnalyzer
 from app.services.nature_resource_analyzer import NatureResourceAnalyzer
-from app.services.rs_service import RemoteSensingService
+from app.services.rs.spectral_engine import spectral_engine
 from app.core.config import settings
 from app.models.upload import UploadRecord
 from app.tools._utils import db_session
